+++ src/solver/conflict.rs
@@ -37,13 +37,13 @@ pub fn handle_conflict(
     let num_conflict = asg.num_conflict;
     // If we can settle this conflict w/o restart, solver will get a big progress.
     let switch_chronobt = if num_conflict < 1000 {
-        Some(false)
+        Some(false) // force normal backtrack
     } else {
-        None
+        None // a closure will determine
     };
     rst.update(ProgressUpdate::Counter);
     // rst.block_restart(); // to update asg progress_evaluator
-    let mut use_chronobt = switch_chronobt.unwrap_or(0 < state.config.c_cbt_thr);
+    let mut use_chronobt = switch_chronobt.unwrap_or_else(|| 0 < state.config.c_cbt_thr);
     if use_chronobt {
         let level = asg.level_ref();
         let c = &mut cdb[ci];
@@ -182,6 +182,7 @@ pub fn handle_conflict(
         return Err(SolverError::NullLearnt);
     }
     // asg.bump_vars(asg, cdb, ci);
+    let chrono_bt_threshold = state.chrono_bt_threshold;
     let new_learnt = &mut state.new_learnt;
     let l0 = new_learnt[0];
     // assert: 0 < cl, which was checked already by new_learnt.is_empty().
@@ -189,11 +190,7 @@ pub fn handle_conflict(
     // NCB places firstUIP on level bl, while CB does it on level cl.
     // Therefore the condition to use CB is: activity(firstUIP) < activity(v(bl)).
     // PREMISE: 0 < bl, because asg.decision_vi accepts only non-zero values.
-    use_chronobt &= switch_chronobt.unwrap_or(
-        bl_a == 0
-            || state.config.c_cbt_thr + bl_a <= cl
-            || asg.activity(l0.vi()) < asg.activity(asg.decision_vi(bl_a)),
-    );
+    use_chronobt &= switch_chronobt.unwrap_or(bl_a == 0 || chrono_bt_threshold + bl_a <= cl);
 
     // (assign level, backtrack level)
     let (al, bl) = if use_chronobt {