+++ ChangeLog.md
@@ -1,5 +1,8 @@
-## 0.6.4, 2021-02-XX
+## 0.6.4, 2021-03-XX
 
+- abolish feature moving_var_reward_rate, switching to stage-based reward decay increment
+- ends of stages trigger in-processors like clause simplifier and clause vivifier
+- Learning Rate based var rewarding uses a non-linear weight function.
 - Since Luby stabilization has long cycles, cycle-based tasks were organized. "Stage" is defined based on Luby-stabilization span. Updated terminology:
    - _Luby stabilization_ is a scheme that defines duration of stabilization modes based on Luby series.
    - Luby stabilization stage, or _stage_ is a span that shares a Luby value.